% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssim_raster.R
\name{ssim_raster}
\alias{ssim_raster}
\title{Calculate SSIM index for raster images}
\usage{
ssim_raster(img1, img2, global = TRUE, w = 3, k1 = NULL, k2 = NULL)
}
\arguments{
\item{img1}{Raster object representing the first image.}

\item{img2}{Raster object representing the second image.}

\item{global}{If global is True, returning the global average of SSIM, SIM, SIV, and SIP. If the option is FALSE, a a \code{raster} raster brick containing the SSIM, SIM, SIV, and SIP for each cell is returned
Default is TRUE}

\item{w}{Integer specifying the window size for the local neighborhood. Default is 3.}

\item{k1}{the constant used in the SSIM calculation. Default is NULL, in which case it is computed from the maximum value of variables.}

\item{k2}{the constant used in the SSIM calculation. Default is NULL, in which case it is computed from the maximum value of variables.}
}
\value{
If global is True, returning the global average of SSIM, SIM, SIV, and SIP. If the option is FALSE, a \code{raster} raster brick containing the SSIM, SIM, SIV, and SIP for each cell is returned
}
\description{
This function calculates the SSIM, a measure of similarity between two raster images
}
\details{
This function computes the SSIM index for two raster images.
}
\examples{
image_path1 <- system.file("data", "groups2nm.tif", package = "SSIMmap")
image_path2 <- system.file("data", "single2nm.tif", package = "SSIMmap")
img1<-rast(image_path1)
img2<-rast(image_path2)
ssim_raster(img1,img2)
result_raster<-ssim_raster(img1,img2,global=FALSE)
result_raster
plot(result_raster)
}
